pyproject.toml
[project]
name = "otel-experiment"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "ipykernel>=7.2.0",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-exporter-otlp-proto-grpc>=1.39.1",
    "opentelemetry-sdk>=1.39.1",
]
